refactor: RacingGame에 전략패턴 도입
