--- src/main/kotlin/racingcar/domain/RandomGenerator.kt
@@ -1,5 +1,0 @@
-package racingcar.domain
-
-interface RandomGenerator {
-    fun getRandomValidation(): Boolean
-}
