+++ src/main/kotlin/racingcar/domain/racing_game/RacingGame.kt
@@ -1,0 +1,8 @@
+package racingcar.domain.racing_game
+
+import racingcar.model.Car
+
+interface RacingGame {
+    val cars: List<Car>
+    fun race()
+}
