+++ src/test/kotlin/racingcar/domain/racing_game/FakeRandomGenerator.kt
@@ -1,4 +1,6 @@
-package racingcar.domain
+package racingcar.domain.racing_game
+
+import racingcar.domain.random.RandomGenerator
 
 class FakeTrueGenerator : RandomGenerator {
     override fun getRandomValidation(): Boolean = true
